일정관리 편집 기능 구현
장소 삭제된 marker 제거 및 해당 일자에 해당하는 marker와 경로 구현, 편집 시 해당하는 DB에 Update, Delete 기능 추가
