+++ src/main/reactapp/package.json
@@ -25,6 +25,7 @@
     "openai": "^4.77.0",
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
+    "react-beautiful-dnd": "^13.1.1",
     "react-bootstrap": "^2.10.5",
     "react-chartjs-2": "^5.2.0",
     "react-datepicker": "^7.5.0",
