+++ src/main/reactapp/src/pages/schedule/scss/ScheduleDate.scss
@@ -1,0 +1,150 @@
+// 변수 정의
+$primary-color: #5D2FFF;
+$border-color: #dee2e6;
+$gray-text: #6c757d;
+
+// 달력 스타일 커스터마이징
+.react-datepicker {
+  font-family: inherit !important;
+  border: none !important;
+  width: 100%;
+  
+  &__header {
+    background-color: white !important;
+    border-bottom: none !important;
+    padding-top: 0 !important;
+  }
+  
+  &__month {
+    margin: 1rem !important;
+  }
+  
+  &__day-names {
+    margin-top: 1rem;
+  }
+  
+  &__day-name {
+    color: $gray-text !important;
+    font-size: 0.875rem;
+    margin: 0.5rem !important;
+    width: 2rem !important;
+    line-height: 2rem !important;
+  }
+  
+  &__day {
+    margin: 0.5rem !important;
+    width: 2rem !important;
+    line-height: 2rem !important;
+    border-radius: 50% !important;
+    color: #333 !important;
+    
+    &:hover {
+      background-color: rgba($primary-color, 0.1) !important;
+    }
+    
+    &--selected {
+      background-color: $primary-color !important;
+      color: white !important;
+      
+      &:hover {
+        background-color: darken($primary-color, 5%) !important;
+      }
+    }
+    
+    &--keyboard-selected {
+      background-color: rgba($primary-color, 0.1) !important;
+      color: #333 !important;
+    }
+    
+    &--in-range {
+      background-color: rgba($primary-color, 0.1) !important;
+      color: #333 !important;
+    }
+    
+    &--in-selecting-range {
+      background-color: rgba($primary-color, 0.1) !important;
+      color: #333 !important;
+    }
+    
+    &--disabled {
+      color: #ccc !important;
+      
+      &:hover {
+        background-color: transparent !important;
+      }
+    }
+  }
+  
+  &__current-month {
+    display: none;
+  }
+  
+  &__navigation {
+    top: 1rem;
+    
+    &--previous {
+      left: 1rem;
+    }
+    
+    &--next {
+      right: 1rem;
+    }
+  }
+}
+
+// 날짜 선택 영역 스타일
+.date-selection-summary {
+  background-color: #f8f9fa;
+  border-radius: 8px;
+  padding: 1rem;
+  margin-top: 1rem;
+  
+  .date-label {
+    font-size: 0.875rem;
+    color: $gray-text;
+    margin-bottom: 0.25rem;
+  }
+  
+  .date-value {
+    font-size: 1rem;
+    font-weight: 600;
+    color: $primary-color;
+  }
+  
+  .duration {
+    font-weight: 600;
+    color: $primary-color;
+    text-align: right;
+  }
+}
+
+// 달력 헤더 커스텀 스타일
+.calendar-header {
+  padding: 0.5rem 1rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  
+  .month-year {
+    font-size: 1.1rem;
+    font-weight: 600;
+    color: #333;
+  }
+  
+  .nav-button {
+    background: none;
+    border: none;
+    padding: 0.5rem;
+    color: #333;
+    cursor: pointer;
+    
+    &:disabled {
+      color: #ccc;
+      cursor: not-allowed;
+    }
+    
+    &:hover:not(:disabled) {
+      color: $primary-color;
+    }
+  }
+}
