+++ src/main/reactapp/src/pages/schedule/scss/ScheduleEdit.scss
@@ -1,0 +1,133 @@
+.schedule-edit-container {
+    width: 500px;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    background: #fff;
+    
+    .edit-header {
+        padding: 1rem;
+        border-bottom: 1px solid #dee2e6;
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        
+        h5 {
+            margin: 0;
+        }
+        
+        .btn-group {
+            display: flex;
+            gap: 0.5rem;
+        }
+    }
+    
+    .days-tabs {
+        padding: 1rem;
+        display: flex;
+        gap: 0.5rem;
+        border-bottom: 1px solid #dee2e6;
+        
+        .day-tab {
+            padding: 0.5rem 1rem;
+            border: 1px solid #dee2e6;
+            border-radius: 0.25rem;
+            background: none;
+            
+            &.active {
+                background: var(--bs-primary);
+                color: white;
+                border-color: var(--bs-primary);
+            }
+        }
+    }
+    
+    .places-container {
+        flex: 1;
+        padding: 1rem;
+        overflow-y: auto;
+        
+        .places-list {
+            display: flex;
+            flex-direction: column;
+            gap: 0.5rem;
+            margin-bottom: 1rem;
+        }
+        
+        .place-item {
+            display: flex;
+            align-items: center;
+            justify-content: space-between;
+            padding: 1rem;
+            background: #f8f9fa;
+            border-radius: 0.5rem;
+            border: 1px solid #dee2e6;
+            
+            &:hover {
+                background: #f1f3f5;
+            }
+            
+            .place-content {
+                display: flex;
+                align-items: center;
+                gap: 1rem;
+            }
+            
+            .order-badge {
+                width: 24px;
+                height: 24px;
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                background: var(--bs-primary);
+                color: white;
+                border-radius: 50%;
+                font-size: 0.875rem;
+            }
+        }
+        
+        .add-place-btn {
+            width: 100%;
+            padding: 1rem;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            gap: 0.5rem;
+        }
+    }
+}
+
+.modal {
+    background-color: rgba(0, 0, 0, 0.5);
+}
+
+.modal-dialog {
+    max-width: 800px;
+}
+
+.modal-content {
+    border-radius: 0.5rem;
+    border: none;
+    
+    .modal-header {
+        border-bottom: 1px solid #dee2e6;
+        padding: 1rem;
+        
+        .modal-title {
+            margin: 0;
+            font-size: 1.25rem;
+        }
+    }
+    
+    .modal-body {
+        padding: 0;
+        overflow: hidden;
+    }
+}
+
+.place-selector-container {
+    height: 100%;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+}
