+++ src/main/reactapp/src/pages/schedule/scss/ScheduleUI.scss
@@ -334,4 +334,30 @@ $shadow: rgba(0, 0, 0, 0.1);
         align-items: center;
         justify-content: center;
     }
+}
+
+.edit-content {
+    &::-webkit-scrollbar {
+        width: 8px;
+    }
+    
+    &::-webkit-scrollbar-track {
+        background: #f1f1f1;
+        border-radius: 4px;
+    }
+    
+    &::-webkit-scrollbar-thumb {
+        background: #888;
+        border-radius: 4px;
+        
+        &:hover {
+            background: #555;
+        }
+    }
+}
+
+.modal-content {
+    border-radius: 1rem;
+    border: none;
+    box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
 }
